pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image2pptx"
version = "0.1.0"
description = "Convert flat illustrations into editable PowerPoint decks."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Cursor Agent" }
]
dependencies = [
  "opencv-python",
  "Pillow",
  "numpy",
  "python-pptx",
  "vtracer",
  "rapidocr-onnxruntime",
  "ultralytics",
  "typer>=0.12",
  "pydantic>=2.7",
]

[project.optional-dependencies]
windows = [
  "pywin32",
]
dev = [
  "pytest",
]

[project.scripts]
image2pptx = "image2pptx.cli:app"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
